# docs/openapi.yaml
openapi: 3.1.0
info:
  title: QueryRouter++
  description: Multi-criteria LLM routing API based on formalized query-model compatibility.
  license:
    name: MIT
  version: 0.2.0
paths:
...
  /v1/chat/completions:
    post:
      summary: Chat Completions
      description: 'OpenAI-compatible chat completions with intelligent routing.


        Accepts standard OpenAI chat format. Routes to the best model based

        on the user''s mode/preset, then proxies the request to the provider.


        Supports LibreChat 4-modes preset names (mode-ecologique, mode-performance,
        etc.)'
      operationId: chat_completions_v1_chat_completions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Chat Completions V1 Chat Completions Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ChatCompletionRequest:
      properties:
        model:
          type: string
          title: Model
          default: queryrouter-auto
        messages:
          items:
            $ref: '#/components/schemas/ChatMessage'
          type: array
          title: Messages
        temperature:
          anyOf:
          - type: number
          - type: 'null'
          title: Temperature
        top_p:
          anyOf:
          - type: number
          - type: 'null'
          title: Top P
        max_tokens:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Tokens
        stream:
          type: boolean
          title: Stream
          default: false
        routing_preference:
          anyOf:
          - type: string
          - type: 'null'
          title: Routing Preference
          description: 'Routing preference: performance, cost, cost_performance, ecology,
            balanced'
      additionalProperties: true
      type: object
      required:
      - messages
      title: ChatCompletionRequest
    ChatMessage:
      properties:
        role:
          type: string
          title: Role
        content:
          anyOf:
          - type: string
          - items: {}
            type: array
          - type: 'null'
          title: Content
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
      type: object
      required:
      - role
      title: ChatMessage
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError